Move ModuleViewMixin to init.py so it can be used by both views
